MenuItem as a Link Component - no need for useNavigate
